requirements.txt
fastapi>=0.110.0
uvicorn>=0.23.0
python-dotenv>=1.0.1
pydantic>=2.9.2
python-multipart
requests
google-generativeai>=0.8.5
google-genai>=1.16.0
google-generativeai
httpx